se agregan pruebas unitaras a servicio crearauto

+++ java-arquitectura-hexagonal/microservicio/aplicacion/src/main/java/com/ceiba/auto/comando/ComandoAuto.java
@@ -1,9 +1,5 @@
 package com.ceiba.auto.comando;
 
-
-
-import java.math.BigDecimal;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,11 +9,11 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class ComandoAuto{
+public class ComandoAuto {
 
-    private String placa;
-    private String tipoCombustible;
-    private Boolean rentado;
+	private String placa;
+	private String tipoCombustible;
+	private Boolean rentado;
 	private Double precioPorDia;
-	private BigDecimal multiplicadorFinSemana;
+	private Double multiplicadorFinSemana;
 }
