+++ java-arquitectura-hexagonal/microservicio/infraestructura/src/main/java/com/ceiba/auto/adaptador/dao/MapeoAuto.java
@@ -1,9 +1,7 @@
 package com.ceiba.auto.adaptador.dao;
 
-import java.math.BigDecimal;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-
 
 import com.ceiba.infraestructura.jdbc.MapperResult;
 import com.ceiba.auto.modelo.dto.DtoAuto;
@@ -11,17 +9,17 @@ import org.springframework.jdbc.core.RowMapper;
 
 public class MapeoAuto implements RowMapper<DtoAuto>, MapperResult {
 
-    @Override
-    public DtoAuto mapRow(ResultSet resultSet, int rowNum) throws SQLException {
+	@Override
+	public DtoAuto mapRow(ResultSet resultSet, int rowNum) throws SQLException {
 
-        Long id = resultSet.getLong("id");
-        String placa = resultSet.getString("placa");
-        String tipoCombustible = resultSet.getString("tipoCombustible");
-        Boolean rentado = resultSet.getBoolean("rentado");
-        Double precioPorDia = resultSet.getDouble("precioPorDia");
-        BigDecimal multiplicadorFinSemana = resultSet.getBigDecimal("multiplicadorFinSemana");
+		Long id = resultSet.getLong("id");
+		String placa = resultSet.getString("placa");
+		String tipoCombustible = resultSet.getString("tipoCombustible");
+		Boolean rentado = resultSet.getBoolean("rentado");
+		Double precioPorDia = resultSet.getDouble("precioPorDia");
+		Double multiplicadorFinSemana = resultSet.getDouble("multiplicadorFinSemana");
 
-        return new DtoAuto(id,placa,tipoCombustible,rentado, precioPorDia, multiplicadorFinSemana);
-    }
+		return new DtoAuto(id, placa, tipoCombustible, rentado, precioPorDia, multiplicadorFinSemana);
+	}
 
 }
